Created TagInput abstraction.
