+++ src/components/TagInput.css
@@ -1,0 +1,10 @@
+/* Hides the placeholder text if at least one tag is present */
+.tagify__tag + .tagify__input::before{
+    display:none !important;
+  }
+
+/* Fixes styling issues between the Tagify input and the Bootstrap input style */
+.tagifyBootstrap {
+    line-height: 1 !important;
+    padding: 0 !important;
+}
